+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/components/bankAccounts/compose/BankAccountsView_Loading.kt
@@ -1,4 +1,43 @@
 package app.cybrid.sdkandroid.components.bankAccounts.compose
 
-class BankAccountsView_Loading {
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import app.cybrid.sdkandroid.R
+import app.cybrid.sdkandroid.ui.Theme.robotoFont
+
+@Composable
+fun BankAccountsView_Loading() {
+
+    Box(
+        modifier = Modifier
+            .height(120.dp)
+    ) {
+        Column(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(
+                text = stringResource(id = R.string.bank_accounts_view_loading_text),
+                fontFamily = robotoFont,
+                fontWeight = FontWeight.Normal,
+                fontSize = 17.sp,
+                color = colorResource(id = R.color.primary_color)
+            )
+            CircularProgressIndicator(
+                modifier = Modifier
+                    .padding(top = 16.dp),
+                color = colorResource(id = R.color.primary_color)
+            )
+        }
+    }
 }
